Mocha revamp (#46)
Separate all tests
doc(constructor): partials & intents
perf(test): added argument --exit to mocha
Updated tests
